+++ crates/icp/src/operations/deploy.rs
@@ -146,6 +146,9 @@ pub enum DeployError {
     },
 
     #[snafu(transparent)]
+    LoadProject { source: ProjectLoadError },
+
+    #[snafu(transparent)]
     Sync { source: SyncOperationError },
 }
 
@@ -588,12 +591,14 @@ async fn sync(
         .collect();
 
     let pkg_cache = ctx.dirs.package_cache()?;
+    let project_dir = ctx.project.load().await?.dir;
 
     let phase = reporter.task(Task::phase("Syncing canisters:"));
     let result = sync_many(
         ctx.syncer.clone(),
         agent.clone(),
         sync_canisters,
+        project_dir,
         environment_selection.name().to_owned(),
         env.network.name.clone(),
         canister_ids,